requirements.txt
openai==0.28.1
python-dotenv==1.0.0
better-profanity==0.7.0
click==8.1.3
plotly==5.19.0
